refactor common name fetching into getTaxonCommonName utility; show common names non-species result screen
